+++ lib/sparql_client.ex
@@ -309,6 +309,26 @@ defmodule SPARQL.Client do
     unvalidated_update(form, update, endpoint, opts)
   end
 
+  def load(endpoint, opts) when is_list(opts) do
+    {from, opts} = Keyword.pop!(opts, :from)
+    {to, opts} = Keyword.pop(opts, :to)
+    {silent, opts} = Keyword.pop(opts, :silent)
+
+    with {:ok, update_string} <- Client.Update.Builder.load(from, to, silent) do
+      do_update(:load, update_string, endpoint, opts)
+    end
+  end
+
+  def load(update, endpoint, opts) do
+    if Keyword.has_key?(opts, :from) or Keyword.has_key?(opts, :to) or
+         Keyword.has_key?(opts, :silent) do
+      raise ArgumentError,
+            "load/3 does not support the :from, :to and :silent options; use load/2 instead"
+    end
+
+    update_data(:load, update, endpoint, opts)
+  end
+
   def clear(endpoint, opts) when is_list(opts) do
     {graph, opts} = Keyword.pop!(opts, :graph)
     {silent, opts} = Keyword.pop(opts, :silent)